requirements.txt
numpy>=1.16.2
tensorflow>=2.3.1
pillow>=8.0.0
iptcinfo3>=2.0.0
imgtag
